[FEAT] - Wire signInAction to postSignInRedirect, add member preview routes
Splits each member page into a presentational view component
(HomeView, PoolDetailView, PayView, InboxView, ProfileView) so the
dev-only preview routes can render the same JSX against static
fixtures. Pages stay thin — they fetch data, derive view models, and
hand off to the matching view.

signInAction now sanitises the supplied callbackUrl first; an unsafe
or empty value falls through to postSignInRedirect, which routes
admins with a non-empty receipts queue to /admin/receipts and
everyone else to /home. Slice 2 cannot count pending receipts yet
(no API), so admins land on /home for now; slice 3 wires the count.

Adds /internal/preview/member/{home,pool,pay,inbox,profile}, each
gated to non-production builds via NODE_ENV check + notFound. The
fixtures live at lib/preview/member-fixtures.ts and mirror the
artboard numbers (Lagos Rent Q2, Family group, Ibadan trip,
ChamaSave) so the screenshot matrix matches the design source.

Existing signin-action tests updated: the "external callbackUrl"
case now lands on /home instead of /, reflecting the role-default
landing. Adds two tests covering member + super_admin no-callback
paths.
